Create python package from the launcher
