+++ launcher/common.py
@@ -1,12 +1,17 @@
 import os
+import shutil
 import subprocess
 
 import psutil
 
 module_root = os.path.abspath(os.path.dirname(os.path.realpath(__file__)))
 project_root = os.path.abspath(os.path.join(os.path.dirname(os.path.realpath(__file__)), '..', '..'))
-python_exe = os.path.join(project_root, 'Python', 'python.exe')
-ffmpeg_path = os.path.join(project_root, 'ffmpeg')
+python_exe = os.environ['_']
+if os.name == "nt":
+    python_exe = os.path.join(project_root, 'Python', 'python.exe')
+ffmpeg_path = 'ffmpeg'
+if os.name == "nt":
+    ffmpeg_path = os.path.join(project_root, 'ffmpeg')
 
 
 def exec_process(subprocess_command: list):
